Create dedicated sections for important constructor arguments
In particular, the underlying source API, the underlying sink API, the transformer API, and queuing strategy objects.

This is inspired by some of the discussion in #872. The new locations for these definitions allow us to link to them, and the more spacious format allows us to flesh out more aspects of the behavior of each option. In particular, this tries to clarify how the return value of an underlying sink's write() method is used, per conversations with Lennart Grahl.